Move bundle -> bundles
Fixes #512
